+++ apps/etterlatte-saksbehandling-ui/client/src/shared/api/behandling.ts
@@ -15,7 +15,7 @@ import { FoersteVirk, ISak } from '~shared/types/sak'
 import { InstitusjonsoppholdMedKilde } from '~components/person/uhaandtereHendelser/HistoriskeHendelser'
 import { format } from 'date-fns'
 import { DatoFormat } from '~utils/formattering'
-import { Aldersgruppe, Brevoppsett } from '~components/behandling/brevoppsett/Brevoppsett'
+import { Brevoppsett } from '~components/behandling/brevoppsett/Brevoppsett'
 
 export const hentGrunnlagsendringshendelserForPerson = async (
   fnr: string
@@ -151,14 +151,6 @@ export const lagreBrevoppsett = async (args: {
   return apiClient.post(`/behandling/${args.behandlingId}/brevoppsett`, { ...args.brevoppsett })
 }
 
-export const hentBrevoppsett = async (behandlingId: string): Promise<ApiResponse<Brevoppsett>> => {
-  console.log(behandlingId)
-  return new Promise<ApiResponse<Brevoppsett>>((resolve) => {
-    resolve({
-      ok: true,
-      status: 200,
-      data: { etterbetaling: { fom: new Date(), tom: new Date() }, aldersgruppe: Aldersgruppe.OVER_18 },
-    })
-  })
-  //return apiClient.get(`/behandling/${behandlingId}/brevoppsett`)
+export const hentBrevoppsett = async (behandlingId: string): Promise<ApiResponse<Brevoppsett | null>> => {
+  return apiClient.get(`/behandling/${behandlingId}/brevoppsett`)
 }
